LayoutTests/platform/chromium-mac/svg/dynamic-updates/SVGFEColorMatrixElement-svgdom-values-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'values' property of the SVGFEColorMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS SVGNumberListToString(matrixElement.values.baseVal) is "0.9 0.9 0.9 0 0 0.33 0.33 0.33 0 0 0.33 0.33 0.33 0 0 0.33 0.33 0.33 0 0 "
PASS Math.round(saturateElement.values.baseVal.getItem(0).value * 1000) is 500
PASS Math.round(hueRotateElement.values.baseVal.getItem(0).value * 1000) is 10000
FAIL SVGNumberListToString(matrixElement.values.baseVal) should be 0.33 0.33 0.33 0 0 0.33 0.33 0.33 0 0 0.33 0.33 0.33 0 0 0.33 0.33 0.33 0 0 . Was NaN NaN NaN 0 0 0.33 0.33 0.33 0 0 0.33 0.33 0.33 0 0 0.33 0.33 0.33 0 0 .
FAIL saturateElement.values.baseVal.getItem(0).value should be 0.25. Was NaN.
FAIL hueRotateElement.values.baseVal.getItem(0).value should be 90. Was NaN.
PASS successfullyParsed is true

TEST COMPLETE
